ref: consolidate BuffReport/DebuffReport into alteration-report.ts
Pair 2 of #217: buff-report.ts and debuff-report.ts shared an identical
inner entry type (target, kind, value, remainingTurns). Extracted the
shared AlterationEntry type and consolidated both report exports into
a single alteration-report.ts file.
