+++ src/fight/core/fight-simulator/@types/alteration-report.ts
@@ -2,18 +2,27 @@ import { AlterationType } from '../../cards/@types/alteration/alteration-type';
 import { CardInfo } from '../../cards/@types/card-info';
 import { StepKind } from './step';
 
-type Debuff = {
+type AlterationEntry = {
   target: CardInfo;
   kind: AlterationType;
   value: number;
   remainingTurns: number;
 };
 
+export type BuffReport = {
+  kind: StepKind.Buff;
+  name?: string;
+  source: CardInfo;
+  buffs: AlterationEntry[];
+  energy: number;
+  powerId?: string;
+};
+
 export type DebuffReport = {
   kind: StepKind.Debuff;
   name?: string;
   source: CardInfo;
-  debuffs: Debuff[];
+  debuffs: AlterationEntry[];
   energy: number;
   powerId?: string;
 };
