+++ src/fight/core/fight-simulator/@types/attack-report.ts
@@ -1,6 +1,6 @@
 import { DamageReport } from './damage-report';
 import { Step, StepKind } from './step';
-import { BuffReport } from './buff-report';
+import { BuffReport } from './alteration-report';
 
 export type AttackReport = {
   kind: StepKind.Attack | StepKind.SpecialAttack;
